🚚 Move: 프리셋 api 폴더 구조 수정
