+++ src/services/preset/deleteCategory.ts
