+++ src/services/preset/index.ts
@@ -1,0 +1,8 @@
+export { deleteCategory } from '@/services/preset/deleteCategory';
+export { deletePresetItem } from '@/services/preset/deletePresetItem';
+export { getAllCategoryList } from '@/services/preset/getAllCategoryList';
+export { getAllCategoryKeywordList } from '@/services/preset/getAllCategoryKeywordList';
+export { getAllCategoryName } from '@/services/preset/getAllCategoryName';
+export { getSingleCategoryList } from '@/services/preset/getSingleCategoryList';
+export { postCreateCategory } from '@/services/preset/postCreateCategory';
+export { postCreatePresetItem } from '@/services/preset/postCreatePresetItem';
