--- src/services/setting/preset/index.ts
@@ -1,8 +1,0 @@
-export { deleteCategory } from '@/services/setting/preset/deleteCategory';
-export { deletePresetItem } from '@/services/setting/preset/deletePresetItem';
-export { getAllCategoryList } from '@/services/setting/preset/getAllCategoryList';
-export { getAllCategoryKeywordList } from '@/services/setting/preset/getAllCategoryKeywordList';
-export { getAllCategoryName } from '@/services/setting/preset/getAllCategoryName';
-export { getSingleCategoryList } from '@/services/setting/preset/getSingleCategoryList';
-export { postCreateCategory } from '@/services/setting/preset/postCreateCategory';
-export { postCreatePresetItem } from '@/services/setting/preset/postCreatePresetItem';
